@@ -11,7 +11,7 @@
     }
   ],
   "require": {
-    "php": ">= 7.2",
+    "php": ">= 7.4",
     "ext-json": "*",
     "dhii/services": "^0.1.1-alpha3",
     "dhii/containers": "^v0.1.4",
