@@ -8,11 +8,7 @@ use Psr\Container\ContainerInterface;
 use RebelCode\WpSdk\Handler;
 use RebelCode\WpSdk\Module;
 use RebelCode\WpSdk\Plugin;
-use RebelCode\WpSdk\Wp\AdminMenu;
 use RebelCode\WpSdk\Wp\NoticeManager;
-use RebelCode\WpSdk\Wp\RestEndpoint;
-use RebelCode\WpSdk\Wp\Shortcode;
-use WP_Block_Type;
 
 /**
  * A module that adds base services for WordPress entities and hooks to register them.
@@ -29,70 +25,66 @@ class WordPressModule extends Module
     }
 
     /** @inheritDoc */
-    public function getHooks(): array
+    public function getHooks(): iterable
     {
-        return [
-            'init' => new Handler(
-                ['post_types', 'shortcodes', 'block_types'],
-                function (array $postTypes, array $shortcodes, array $blockTypes) {
-                    foreach ($postTypes as $postType) {
-                        $postType->register();
-                    }
-                    foreach ($shortcodes as $shortcode) {
-                        /** @var $shortcode Shortcode */
-                        $shortcode->register();
-                    }
-                    foreach ($blockTypes as $blockType) {
-                        /** @var $blockType WP_Block_Type */
-                        register_block_type($blockType);
-                    }
+        yield 'init' => new Handler(
+            ['post_types', 'shortcodes', 'block_types'],
+            function (array $postTypes, array $shortcodes, array $blockTypes) {
+                foreach ($postTypes as $postType) {
+                    $postType->register();
                 }
-            ),
+                foreach ($shortcodes as $shortcode) {
+                    /** @var $shortcode Shortcode */
+                    $shortcode->register();
+                }
+                foreach ($blockTypes as $blockType) {
+                    /** @var $blockType WP_Block_Type */
+                    register_block_type($blockType);
+                }
+            }
+        );
 
-            'plugins_loaded' => new Handler(
-                ['cron_jobs'],
-                function (array $cronJobs) {
-                    foreach ($cronJobs as $cronJob) {
-                        $cronJob->registerHandlers();
-                    }
+        yield 'plugins_loaded' => new Handler(
+            ['cron_jobs'],
+            function (array $cronJobs) {
+                foreach ($cronJobs as $cronJob) {
+                    $cronJob->registerHandlers();
                 }
-            ),
+            }
+        );
 
-            'admin_menu' => new Handler(
-                ['admin_menus'],
-                function (array $menus) {
-                    foreach ($menus as $menu) {
-                        /** @var $menu AdminMenu */
-                        $menu->register();
-                    }
+        yield 'admin_menu' => new Handler(
+            ['admin_menus'],
+            function (array $menus) {
+                foreach ($menus as $menu) {
+                    /** @var $menu AdminMenu */
+                    $menu->register();
                 }
-            ),
+            }
+        );
 
-            'rest_api_init' => new Handler(
-                ['rest_endpoints'],
-                function (array $endpoints) {
-                    foreach ($endpoints as $endpoint) {
-                        /** @var $endpoint RestEndpoint */
-                        $endpoint->register();
-                    }
+        yield 'rest_api_init' => new Handler(
+            ['rest_endpoints'],
+            function (array $endpoints) {
+                foreach ($endpoints as $endpoint) {
+                    /** @var $endpoint RestEndpoint */
+                    $endpoint->register();
                 }
-            ),
-        ];
+            }
+        );
     }
 
     /** @inheritDoc */
-    public function getFactories(): array
+    public function getFactories(): iterable
     {
-        return [
-            'db' => new GlobalVar('wpdb'),
-            'post_types' => new Value([]),
-            'notices' => new Value([]),
-            'notices/manager' => NoticeManager::factory('@plugin/short_id', 'notices'),
-            'admin_menus' => new Value([]),
-            'cron_jobs' => new Value([]),
-            'rest_endpoints' => new Value([]),
-            'shortcodes' => new Value([]),
-            'block_types' => new Value([]),
-        ];
+        yield 'db' => new GlobalVar('wpdb');
+        yield 'post_types' => new Value([]);
+        yield 'notices' => new Value([]);
+        yield 'notices/manager' => NoticeManager::factory('@plugin/short_id', 'notices');
+        yield 'admin_menus' => new Value([]);
+        yield 'cron_jobs' => new Value([]);
+        yield 'rest_endpoints' => new Value([]);
+        yield 'shortcodes' => new Value([]);
+        yield 'block_types' => new Value([]);
     }
 }
